Move LLM CLI metadata helpers
